+++ semcore/d3-chart/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.4.9] - 2022-11-03
+
+### Fixed
+
+- Fixed hover and active animated for `Donut` chart.
+
 ## [2.4.8] - 2022-11-02
 
 ### Fixed