+++ semcore/date-picker/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [3.4.5] - 2022-10-28
+## [3.4.6] - 2022-11-03
 
 ### Changed
 
-- Version patch update due to children dependencies update (`@semcore/utils` [3.40.0 ~> 3.40.0], `@semcore/base-trigger` [3.2.4 ~> 3.2.5], `@semcore/button` [4.2.2 ~> 4.2.3], `@semcore/icon` [3.1.1 ~> 3.1.2]).
+- Version patch update due to children dependencies update (`@semcore/input-mask` [4.3.3 ~> 4.3.4]).
 
 ## [3.4.1] - 2022-10-17